+++ packages/server/src/api/controllers/ai/sharepointAuth.ts
@@ -1,25 +1,6 @@
-import {
-  cache,
-  configs,
-  constants,
-  context,
-  encryption,
-  env,
-  utils,
-} from "@budibase/backend-core"
-import {
-  Datasource,
-  DatasourceAuthCookie,
-  OAuth2CredentialsMethod,
-  OAuth2GrantType,
-  OAuth2RestAuthConfig,
-  RestAuthType,
-  SourceName,
-  UserCtx,
-  isOAuth2DelegatedAuthConfig,
-} from "@budibase/types"
-import sdk from "../../../sdk"
-import { saveDelegatedOAuthCredential } from "../../../sdk/workspace/ai/knowledgeSources/sharepoint/credentials"
+import { cache, configs, constants, env, utils } from "@budibase/backend-core"
+import { DatasourceAuthCookie, UserCtx } from "@budibase/types"
+import { upsertDelegatedSharePointAuthConfig } from "../../../sdk/workspace/ai/knowledgeSources/sharepoint/authConfig"
 
 const DEFAULT_SCOPE = env.RAG_SHAREPOINT_DEFAULT_SCOPE
 const STATE_CACHE_TTL_SECONDS = 600
@@ -40,17 +21,6 @@ interface SharePointOAuthState {
   provider: string
   datasourceId: string
   authConfigId?: string
-}
-
-interface DelegatedSharePointCredentials {
-  account: string
-  accessToken: string
-  refreshToken: string
-  tokenType: string
-  expiresAt: number
-  tokenEndpoint: string
-  clientId: string
-  clientSecret: string
 }
 
 const getMicrosoftConfig = () => {
@@ -133,91 +103,6 @@ export async function startSharePointAuth(ctx: UserCtx<void, void>) {
   authorizeUrl.searchParams.set("state", state)
 
   ctx.redirect(authorizeUrl.toString())
-}
-
-const isSharePointDatasource = (datasource: Datasource) => {
-  return (
-    datasource.source === SourceName.REST &&
-    (datasource.restTemplateId === "microsoft-sharepoint" ||
-      datasource.restTemplate === "Microsoft SharePoint")
-  )
-}
-
-export const upsertDelegatedSharePointAuthConfig = async (
-  appId: string,
-  datasourceId: string,
-  authConfigId: string | undefined,
-  credentials: DelegatedSharePointCredentials
-) => {
-  return context.doInWorkspaceContext(appId, async () => {
-    const account = credentials.account.trim() || "unknown"
-    const datasource = await sdk.datasources.get(datasourceId)
-    if (!isSharePointDatasource(datasource)) {
-      throw new Error(
-        "SharePoint OAuth can only be used with SharePoint datasources"
-      )
-    }
-
-    const authConfigs = (
-      (datasource.config?.authConfigs || []) as OAuth2RestAuthConfig[]
-    ).filter(Boolean)
-
-    const explicitConfig = authConfigId
-      ? authConfigs.find(config => config._id === authConfigId)
-      : undefined
-    if (authConfigId && !explicitConfig) {
-      throw new Error("SharePoint auth config not found on datasource")
-    }
-
-    const matchingConfig =
-      explicitConfig ||
-      authConfigs.find(
-        config =>
-          isOAuth2DelegatedAuthConfig(config) &&
-          config.account?.toLowerCase() === account.toLowerCase()
-      )
-    const nextAuthConfig: OAuth2RestAuthConfig = {
-      ...(matchingConfig || {}),
-      _id: matchingConfig?._id || `auth_${utils.newid()}`,
-      type: RestAuthType.OAUTH2,
-      authType: "delegated_oauth",
-      name: matchingConfig?.name || `Microsoft SharePoint (${account})`,
-      account,
-      url: credentials.tokenEndpoint,
-      clientId: credentials.clientId,
-      clientSecret: encryption.encrypt(credentials.clientSecret),
-      method: OAuth2CredentialsMethod.BODY,
-      grantType: OAuth2GrantType.CLIENT_CREDENTIALS,
-      scope: DEFAULT_SCOPE,
-    }
-    const nextAuthConfigs = matchingConfig
-      ? authConfigs.map(config =>
-          config._id === matchingConfig._id ? nextAuthConfig : config
-        )
-      : [...authConfigs, nextAuthConfig]
-
-    await sdk.datasources.save({
-      ...datasource,
-      config: {
-        ...datasource.config,
-        authConfigs: nextAuthConfigs,
-      },
-    })
-
-    await saveDelegatedOAuthCredential({
-      datasourceId,
-      authConfigId: nextAuthConfig._id,
-      accessToken: credentials.accessToken,
-      refreshToken: credentials.refreshToken,
-      tokenType: credentials.tokenType,
-      expiresAt: credentials.expiresAt,
-    })
-
-    return {
-      authConfigId: nextAuthConfig._id,
-      reusedExistingConnection: !!matchingConfig,
-    }
-  })
 }
 
 export async function completeSharePointAuth(ctx: UserCtx<void, void>) {
@@ -350,7 +235,8 @@ export async function completeSharePointAuth(ctx: UserCtx<void, void>) {
         clientId,
         clientSecret,
         expiresAt: calculateBufferedTokenExpiry(expiresIn),
-      }
+      },
+      DEFAULT_SCOPE
     )
   console.log("SharePoint delegated OAuth callback received", {
     appId,
